+++ vote/changelog.md
@@ -2,6 +2,7 @@
 
 - Added support for optional `candidates` in `IrvElection` and 
   `CondorcetElection`.
+- Added `bool get hasSingleWinner` to `Election`.
 
 # 0.1.0+2