+++ vote/changelog.md
@@ -4,6 +4,7 @@
 - Added support for optional `candidates` in `IrvElection` and 
   `CondorcetElection`.
 - Added `bool get hasSingleWinner` to `Election`.
+- Require Dart `2.4.0`.
 
 # 0.1.0+2